Removing Hide for insert context method as we do not need to hide the next item when this method is selected.
This contrib module has been extended to add a class based on the dfp slot trafficked to the page as well as a default class.

It has also retained it’s hide functionality for Replace context methods.

The hope being that out of the box, it works without any site level changes needed. But there are unique classes available for site level CSS if needed.